Added bindings for FormatDocument and FormatSelection.
